requirements.txt
streamlit
langchain-ollama
langchain
langchain-community
pdftext
ollama
pymongo
python-dotenv
uuid
easyocr
pillow
neo4j
ollama-python
requests 
langchain-chroma